Remove all the array schedule gunge.
After thinking about it for a little, we can actually model what we want
straight in SMT using the 'distinct' function. All we want from a round
is that teams only occur once, which can be done in one function
application. All other properties follow from the round-based nature
anyway.