Start experimenting with some quantifier goo.
So: According to everything I read, while AUFLIRA and other such logics
are generally undecideable, there are a lot of solvers out there that
can remove a lot of quantifiers. So we can at least give it a shot, and
see how fast such a thing performs.

What this means is that all kind of explicit things said in the past,
like "these slots in this round are distinct" can be said in one
assertion by quantifying it over things.

Now, this may very well kill performance massively... but it's going to
allow the most compact description of our problem, which means the
solver can optimise away as much state space as possible. Which, in
reality, is what's taking up all the time.

An unanswered question is how one extracts a schedule out of the model
once a satisfying function has been found, however using get-model
returns the formula that makes up that function, so it shouldn't be
difficult. May have to use incremental SMT to extract some numbers.